+++ Project.Data/Entities/CaseType.cs
@@ -3,7 +3,7 @@ using Moujam.Casiher.Comman.Base;
 
 namespace Ettad.Data.Entities
 {
-    public class CaseType : AuditEntity<int>, ILookup
+    public class CaseType : AuditEntity<long>, ILookup
     {
         public string NameAr { get; set; }
         public string NameEn { get; set; }
