+++ Project.Data/Entities/NatureOption.cs
@@ -3,7 +3,7 @@ using Moujam.Casiher.Comman.Base;
 
 namespace Ettad.Data.Entities
 {
-    public class NatureOption : AuditEntity<int>, ILookup
+    public class NatureOption : AuditEntity<long>, ILookup
     {
         public string NameAr { get; set; }
         public string NameEn { get; set; }
