+++ Project.Data/Entities/Nsn.cs
@@ -4,7 +4,7 @@ using Moujam.Casiher.Comman.Base;
 
 namespace Ettad.Data.Entities
 {
-    public class Nsn : AuditEntity<int>, ILookup
+    public class Nsn : AuditEntity<long>, ILookup
     {
         public string NameAr { get; set; }
         public string NameEn { get; set; }
