+++ Project.Data/Entities/Supplier.cs
@@ -3,7 +3,7 @@ using Moujam.Casiher.Comman.Base;
 
 namespace Ettad.Data.Entities
 {
-    public class Supplier : AuditEntity<int>, INameable
+    public class Supplier : AuditEntity<long>, INameable
     {
         public string NameAr { get; set; }
         public string NameEn { get; set; }
